+++ internal/renderers/software/lib.rs
@@ -992,7 +992,7 @@ impl RendererSealed for SoftwareRenderer {
             }
             #[cfg(feature = "systemfonts")]
             (fonts::Font::VectorFont(vf), true) => {
-                let visual_representation = text_input.visual_representation(None);
+                let visual_representation = text_input.visual_representation();
 
                 let width = (text_input.width().cast() * scale_factor).cast();
                 let height = (text_input.height().cast() * scale_factor).cast();
@@ -1020,7 +1020,7 @@ impl RendererSealed for SoftwareRenderer {
                 )
             }
             (fonts::Font::PixelFont(pf), _) => {
-                let visual_representation = text_input.visual_representation(None);
+                let visual_representation = text_input.visual_representation();
 
                 let width = (text_input.width().cast() * scale_factor).cast();
                 let height = (text_input.height().cast() * scale_factor).cast();
@@ -1087,7 +1087,7 @@ impl RendererSealed for SoftwareRenderer {
             }
             #[cfg(feature = "systemfonts")]
             (fonts::Font::VectorFont(vf), true) => {
-                let visual_representation = text_input.visual_representation(None);
+                let visual_representation = text_input.visual_representation();
 
                 let width = (text_input.width().cast() * scale_factor).cast();
                 let height = (text_input.height().cast() * scale_factor).cast();
@@ -1121,7 +1121,7 @@ impl RendererSealed for SoftwareRenderer {
                     .cast()
             }
             (fonts::Font::PixelFont(pf), _) => {
-                let visual_representation = text_input.visual_representation(None);
+                let visual_representation = text_input.visual_representation();
 
                 let width = (text_input.width().cast() * scale_factor).cast();
                 let height = (text_input.height().cast() * scale_factor).cast();
@@ -2828,7 +2828,7 @@ impl<T: ProcessScene> i_slint_core::item_rendering::ItemRenderer for SceneBuilde
             #[cfg(feature = "systemfonts")]
             (fonts::Font::VectorFont(_), false) => {
                 drop(font_ctx);
-                sharedparley::draw_text_input(self, text_input, self_rc, size, None);
+                sharedparley::draw_text_input(self, text_input, self_rc, size);
             }
             #[cfg(feature = "systemfonts")]
             (fonts::Font::VectorFont(vf), true) => {
@@ -2850,7 +2850,7 @@ impl<T: ProcessScene> i_slint_core::item_rendering::ItemRenderer for SceneBuilde
                     };
                 let offset = self.current_state.offset.to_vector().cast() * self.scale_factor;
 
-                let text_visual_representation = text_input.visual_representation(None);
+                let text_visual_representation = text_input.visual_representation();
                 let color = self.alpha_color(text_visual_representation.text_color.color());
 
                 let selection = (!text_visual_representation.selection_range.is_empty()).then_some(
@@ -2934,7 +2934,7 @@ impl<T: ProcessScene> i_slint_core::item_rendering::ItemRenderer for SceneBuilde
                     };
                 let offset = self.current_state.offset.to_vector().cast() * self.scale_factor;
 
-                let text_visual_representation = text_input.visual_representation(None);
+                let text_visual_representation = text_input.visual_representation();
                 let color = self.alpha_color(text_visual_representation.text_color.color());
 
                 let selection = (!text_visual_representation.selection_range.is_empty()).then_some(